feat(tools): unify monitoring into a single `marutop` CLI
Add a new `maru_tools` package and a `marutop` console script, folding the four standalone tools/*.py monitors into
one umbrella command. The old scripts become thin back-compat shims.

- marutop (default): htop-style curses TUI fusing three sections —
  DEVICES (DAX pools from the Resource Manager), INSTANCES (per-instance
  allocated/used/slack from MaruServer), and STATS (compact per-op summary).
  * background poller thread keeps the UI smooth / keys responsive even
    when a server is slow or unreachable (per-section "(unavailable)")
  * maru-server auto-discovery by scanning /proc for their --port, so
    multiple servers are covered without knowing ports (-p/--host pins one)
  * per-instance per-DAX-device gauges (which device each instance landed
    on and how much of it it holds)
  * colored gauges + [s]ort/[i]nterval/[q]uit keys; --once for plain text
- subcommands: marutop pool | usage | stats | device (init|show|clear)
- RM stays a fixed well-known address (:9850, --address to override) since
  it is a singleton daemon; only MaruServers vary and are discovered.

protocol: InstanceUsage gains a `devices` map (dax_path -> bytes), resolved
server-side via MaruShmClient.get_dax_path; backward compatible (defaults {}).

pyproject: register `marutop = maru_tools.cli:main` and package maru_tools.
docs: rewrite tools/README + README around marutop. tests updated (import
from maru_tools.usage; MockShmClient.get_dax_path; devices-breakdown test).
