+++ maru_tools/__init__.py
@@ -1,0 +1,11 @@
+"""Maru monitoring and admin tooling.
+
+A single ``marutop`` command umbrella (see :mod:`maru_tools.cli`) that fuses the
+former standalone ``tools/*.py`` scripts:
+
+- ``marutop``          — unified live pool + per-instance view (default)
+- ``marutop pool``     — physical DAX pool gauges (Resource Manager)
+- ``marutop usage``    — per-instance allocated/used/slack (MaruServer)
+- ``marutop stats``    — per-operation latency/throughput dashboard (MaruServer)
+- ``marutop device``   — DAX device UUID header init/clear/show (Resource Manager)
+"""
